CSS Sprites logic has been fully re-written. Now defined 13 cases: 1. No-repeat w/ height and width (position not in % or ems) 2. No-repeat w/o height or width (position not in % or ems) 3. Repeat-x (w/ height) 4. Repeat-x (w/o height) 5. Repeat-y (w/ width) 6. Repeat-y (w/o width) 7. No-repeat w/ width and position 0 100% 8. No-repeat w/ height and position 100% 0 9. No-repeat w/o width and position 0 100% 10. No-repeat w/o height and position 100% 0 11. No-repeat w/ position center / 100% 100% / etc. 12. Repeat 13. Animated GIFs
First 8 are operable and can be safely merged into 3 CSS Sprites (1 + 2 + 7 + 8 , 1 + 3 + 4, 1 + 5 + 6). In some cases we can aggressively merge 3 + 4 and 5 + 6 -- there will an option in configuration.

Additionally 1 + 2 are placed with optimum minimum of white space (algorithm is not good, not bad). Some very rare and specific cases are not handled -- need workaround.

Also some notices should be fixed in the future -- minor issues.